requirements.txt
Pillow==7.0.0
pika==1.1.0